divided main.py and cleanup code
